Fix Erlang LS dependency resolution for rebar3
This commit adds rebar3 as a project plugin
and updates erlang_ls.yaml so that "go to definition"
and similar features in Erlang LS work correctly for
functions exported by rebar3.

Since rebar3 is a project plugin, it will not be
needlessly downloaded when users add rebar3_grisp as a plugin.
It's only fetched when rebar3_grisp is compiled directly.